pyproject.toml
[tool.poetry]
name = "ingramdocai"
version = "0.1.0"
description = "Agentic AI system for intelligent document processing and analysis"
authors = ["Gabriel Ohaike <gabrielohaike@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
boto3 = "^1.38.15"
pydantic = "^2.11.4"
pandas = "^2.2.3"
openai = "^1.75.0"
requests = "^2.32.3"
python-dotenv = "^1.1.0"
tenacity = "^9.1.2"
schedule = "^1.2.2"
orjson = "^3.10.18"
cryptography = "^44.0.3"
weaviate-client = "^4.14.3"
langchain = "^0.3.25"
psycopg2-binary = "^2.9.10"
crewai = "^0.126.0"
sqlalchemy = "^2.0.41"
langchain-community = "^0.3.27"
unstructured = "^0.18.5"
pymupdf = "^1.26.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
black = "^25.1.0"
isort = "^6.0.1"
mypy = "^1.15.0"
notebook = "^7.4.2"
ipykernel = "^6.29.5"
git-filter-repo = "^2.47.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

packages = [
  { include = "ingramdocai" }
]

[tool.poetry.scripts]
ingramdocai_start = "ingramdocai.main:start"
ingramdocai_plot  = "ingramdocai.main:plot"
